+++ tests/Unit/SqlLoggerTest.php
@@ -34,8 +34,8 @@ class SqlLoggerTest extends UnitTestCase
         ]);
 
         $sqlQuery = new SqlQuery(1, 'anything', [], 1.23);
-//        $this->writer->shouldReceive('save')->once()->with($sqlQuery)
-        $this->writer->shouldReceive('save')->once()->with(Mockery::on(function($arg) use ($sqlQuery) {
+//        $this->writer->shouldReceive('writeQuery')->once()->with($sqlQuery)
+        $this->writer->shouldReceive('writeQuery')->once()->with(Mockery::on(function($arg) use ($sqlQuery) {
             return $sqlQuery == $arg;
         }));
 
@@ -52,14 +52,14 @@ class SqlLoggerTest extends UnitTestCase
         ]);
 
         $sqlQuery = new SqlQuery(1, 'anything', ['one', 1], 1.23);
-//        $this->writer->shouldReceive('save')->once()->with($sqlQuery);
-        $this->writer->shouldReceive('save')->once()->with(Mockery::on(function($arg) use ($sqlQuery) {
+//        $this->writer->shouldReceive('writeQuery')->once()->with($sqlQuery);
+        $this->writer->shouldReceive('writeQuery')->once()->with(Mockery::on(function($arg) use ($sqlQuery) {
             return $sqlQuery == $arg;
         }));
 
         $sqlQuery2 = new SqlQuery(2, 'anything2', ['two', 2], 4.56);
-//        $this->writer->shouldReceive('save')->once()->with($sqlQuery2);
-        $this->writer->shouldReceive('save')->once()->with(Mockery::on(function($arg) use ($sqlQuery2) {
+//        $this->writer->shouldReceive('writeQuery')->once()->with($sqlQuery2);
+        $this->writer->shouldReceive('writeQuery')->once()->with(Mockery::on(function($arg) use ($sqlQuery2) {
             return $sqlQuery2 == $arg;
         }));